// src/utils/types.ts
export type Product = {
  id: number;
  title: string;
  discountPercentage: number;
  price: number;
  stock: number;
  rating: number;
  thumbnail: string;
  description: string;
};

export type ProductL = {
  id: number;
  title: string;
  price: number;
  rating: number;
  description: string;
};